requirements.txt
aistudio-sdk==0.3.8
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
av==16.0.1
bce-python-sdk==0.9.56
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
ctranslate2==4.6.2
faiss-cpu==1.13.1
faster-whisper==1.2.1
filelock==3.20.1
flatbuffers==25.9.23
fsspec==2025.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
modelscope==1.33.0
mpmath==1.3.0
narwhals==2.14.0
networkx==3.6.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.11
pandas==2.3.3
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pypdfium2==5.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
wcwidth==0.2.14
